moves context dependency to prevent circular dependency
