+++ prompting/tools/datasets/__init__.py
@@ -1,4 +1,3 @@
-from .context import Context
 from .base import Dataset
 from .code import HFCodingDataset, StackOverflowDataset
 from .math import MathDataset
